+++ field/field-campaign-code.js
@@ -8,7 +8,7 @@ class FieldCampaignCode extends FieldCode {
   constructor(options = {}) {
     super(options);
     this.lookupFunctionName = 'campaignCode';
-    this.baseTypeId = 10; // TODO: what is the base code for the campaigns??????
+    this.baseTypeId = 62; // TODO: what is the base code for the campaigns??????
   }
 }
 module.exports.FieldCampaignCode = FieldCampaignCode;